views/widgets: fold GTK3-ported demo runners into PR-0's renamed files
PR-0 already renamed test_selectdialog.py/test_selectview.py/
test_textbox.py to demo_selectdialog.py/demo_selectview.py/
demo_textbox.py (pytest was picking them up by filename convention
alone despite having no assertions). This branch's own port work
predates that rename and reconstructed the GTK3-ported content under
the old names, leaving both the old (ported) and new (unported,
inherited from main) files present side by side.

Keep the ported content, under the names PR-0 already settled on.
